requirements.txt
Django>=4.0
psycopg2>=2.8
djangorestframework==3.13.1
pylint
pylint-django
black==22.3.0
whitenoise==6.0.0
gunicorn
django-heroku
